fix: text responses
